refactor: split up UserMenu to parts
