+++ apps/web/components/buttons/FloatingUserButton/UserMenu.tsx
@@ -1,35 +1,7 @@
 import { Menu, Transition } from '@headlessui/react'
-import Link from 'next/link'
-import { FC, ForwardedRef, forwardRef, Fragment, ReactElement } from 'react'
+import { FC, Fragment, ReactElement } from 'react'
 
-function getMenuLink() {
-  const MenuLink = forwardRef(
-    (
-      props: {
-        href: string
-        children: string
-        className: string
-        onClick: () => void
-      },
-      ref: ForwardedRef<HTMLAnchorElement>
-    ) => {
-      const { href, children, className, onClick } = props
-      return (
-        <Link href={href} passHref>
-          <a className={className} onClick={onClick} ref={ref}>
-            {children}
-          </a>
-        </Link>
-      )
-    }
-  )
-  MenuLink.displayName = 'MenuLink'
-  return MenuLink
-}
-
-export const MenuLink = getMenuLink()
-
-export default function DropdownMenu(props: {
+export default function UserMenu(props: {
   TriggerButton: FC
   children: ReactElement[] | ReactElement
 }) {
